app/src/main/java/com/android/movieapp/data/api/MovieService.kt
package com.android.movieapp.data.api

import com.android.movieapp.data.db.MovieEntity
import com.android.movieapp.data.model.Movie
import retrofit2.http.GET
import retrofit2.http.Path
import retrofit2.http.Query


interface MovieService {

    @GET("movie/top_rated")
    suspend fun getMovieList(
        @Query("api_key") apiKey: String,
        @Query("page") page: Int
    ): FeedMoviesResponse<Movie>

    @GET("movie/{movie_id}")
    suspend fun getMovieDetails(
        @Path("movie_id") id: Int,
        @Query("api_key") apiKey: String
    ): Movie
}